Log in now directs user to homepage after successful login
Crazy Swift bug made me spend a week on this. After reading though the
debugging I noticed the keyboard was still open during transitioning,
and by removing the keyboard it actually works. But if the keyboard is
still open the app will still crash. Further improvement needed but
technically works.